requirements.txt
discord.py
sqlmodel
aiosqlite
orjson
python-dotenv
pyhumps